# task-api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "FlyRank Internship Assignment"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],

  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },

  "paths": {
    "/": {
      "get": {
        "summary": "Root endpoint",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },

    "/health": {
      "get": {
        "summary": "Health Check",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },

    "/public/info": {
      "get": {
        "summary": "Public Information",
        "responses": {
          "200": {
            "description": "Public response"
          }
        }
      }
    },

    "/auth/signup": {
      "post": {
        "summary": "Signup User",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User Created"
          },
          "400": {
            "description": "Invalid Input"
          }
        }
      }
    },

    "/auth/login": {
      "post": {
        "summary": "Login User",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login Successful"
          },
          "401": {
            "description": "Invalid Credentials"
          }
        }
      }
    },

    "/auth/logout": {
      "post": {
        "summary": "Logout User",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Logged Out"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },

    "/protected/profile": {
      "get": {
        "summary": "Get User Profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile Returned"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },

    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "responses": {
          "200": {
            "description": "Task List"
          }
        }
      },
      "post": {
        "summary": "Create Task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Learn Express"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task Created"
          }
        }
      }
    },

    "/tasks/{id}": {
      "get": {
        "summary": "Get Task By ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Task Not Found"
          }
        }
      },

      "put": {
        "summary": "Update Task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Learn Node.js"
                  },
                  "done": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task Updated"
          },
          "404": {
            "description": "Task Not Found"
          }
        }
      },

      "delete": {
        "summary": "Delete Task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task Deleted"
          },
          "404": {
            "description": "Task Not Found"
          }
        }
      }
    }
  }
}